+++ FScriptFramework/FSCNAssignment.m
@@ -45,7 +45,7 @@
   return self;
 }
 
-- (void)translateCharRange:(long)translation
+- (void)translateCharRange:(int32_t)translation
 {
   [super translateCharRange:translation];
   [left translateCharRange:translation];
